refactor(meeting_backend/persist): split into mod + cognitive + extract + markdown + templates
persist.rs (1075) → persist/{mod.rs (249), cognitive.rs (149), extract.rs (399),
markdown.rs (255), templates.rs (72)}. All 81 meeting_backend tests pass.

Refs #1266
